fix slice transform
