+++ rocks-sys/src/lib.rs
@@ -1,6 +1,7 @@
 pub use c::*;
 
 pub mod comparator;
+pub mod slice_transform;
 
 #[allow(non_upper_case_globals)]
 #[allow(non_camel_case_types)]
